+++ priv/repo/migrations/20250807102742_rename_relations_table_to_relations_backup.exs
@@ -1,0 +1,7 @@
+defmodule TdLm.Repo.Migrations.AddColumnToRelationsTable do
+  use Ecto.Migration
+
+  def change do
+    rename(table(:relations), to: table(:relations_backup))
+  end
+end